requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-cleanup==2.1.0
django-cors-headers==2.1.0
django-daterange-filter==1.3.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.7.7
drf-nested-routers==0.90.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
libgravatar==0.2.2
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
Pillow==5.0.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
